Move env from 'utils/addresses.py' -> 'service_registry.py'
