+++ backend_services/api_gateway/app/main.py
@@ -12,7 +12,7 @@ from fastapi.middleware.cors import CORSMiddleware
 from api_models.users import UserLoginResponse
 from service_registry import service_registry
 import websockets.client
-from utils.addresses import MATCHING_SERVICE_HOST, COLLABORATION_SERVICE_HOST, COMMUNICATION_SERVICE_HOST
+from service_registry import MATCHING_SERVICE_HOST, COLLABORATION_SERVICE_HOST, COMMUNICATION_SERVICE_HOST
 import websockets.exceptions
 
 app = FastAPI()
